seperated file system
