+++ internal/domain/file.go
@@ -6,18 +6,6 @@ import (
 	"github.com/Rhaqim/buckt/request"
 )
 
-type StorageFileService interface {
-	CreateOwner(name, email string) error
-	CreateBucket(name, description, ownerID string) error
-	UploadFile(file *multipart.FileHeader, bucketName string, folderPath string) error
-	DownloadFile(filename string) ([]byte, error)
-	DeleteFile(filename string) error
-	GetFiles(bucketName string) ([]interface{}, error)
-	Serve(filename string, serve bool) (string, error)
-	GetFilesInFolder(bucketName, folderPath string) ([]interface{}, error)
-	GetSubFolders(bucketName, folderPath string) ([]interface{}, error)
-}
-
 type ManagerService interface {
 	CreateOwner(name, email string) error
 	CreateBucket(name, description, ownerID string) error
