+++ internal/error/error.go
@@ -8,4 +8,5 @@ var (
 	ErrFileAlreadyExists = errors.New("file already exists")
 	ErrBucketNotFound    = errors.New("bucket not found")
 	ErrInvalidUUID       = errors.New("invalid UUID")
+	ErrFileNotFound      = errors.New("file not found")
 )
